requirements.txt
Django==3.1
gunicorn==20.0.4
psycopg2==2.8.5
django-heroku==0.3.1
djangorestframework==3.11.1
